refactor: split colorExtractor into extraction algorithm and cache module
